pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "causal-engine"
version = "0.1.0"
description = "A causal modeling engine for effect estimation and policy simulation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Causal Engine Team"}
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "dowhy>=0.11.0",
    "econml>=0.14.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.7.0",
    "networkx>=3.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydot<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
pdf = [
    "pypdf>=3.0.0",
]

[project.scripts]
causal-engine = "src.api:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

